requirements.txt
Flask
Flask-Cors
sqlalchemy
mysql-connector-python
openai
python-dotenv
